+++ lib/ui/academics/EssentialSkillsCoach.dart
@@ -13,25 +13,30 @@
 // limitations under the License.
 
 import 'package:flutter/cupertino.dart';
-import 'package:illinois/service/Config.dart';
-import 'package:illinois/ui/widgets/AccessWidgets.dart';
+import 'package:flutter/material.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
+import 'package:rokwire_plugin/service/surveys.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 import 'package:rokwire_plugin/ui/widgets/section_header.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
+import 'EssentialSkillsCoachGetStarted.dart';
+import 'EssentialSkillsResults.dart';
+
 
 class EssentialSkillsCoach extends StatefulWidget {
-  final Function onStartCourse;
+  final Function (String?)? onStartCourse;
 
-  EssentialSkillsCoach({required this.onStartCourse});
+  EssentialSkillsCoach({this.onStartCourse});
 
   @override
   _EssentialSkillsCoachState createState() => _EssentialSkillsCoachState();
 }
 
 class _EssentialSkillsCoachState extends State<EssentialSkillsCoach> {
+
+  bool whichPage = true;
 
   @override
   Widget build(BuildContext context) {
@@ -56,7 +61,7 @@ class _EssentialSkillsCoachState extends State<EssentialSkillsCoach> {
           label: Localization().getStringEx("panel.essential_skills_coach.get_started.button.label", 'Get Started'),
           textStyle: Styles().textStyles.getTextStyle("widget.button.title.large.fat.variant"),
           backgroundColor: Styles().colors.surface,
-          onTap: _onTapStartSkillsCoach
+          onTap:_loadResults
         )),
       ]),),
       decoration: BoxDecoration(
@@ -75,7 +80,9 @@ class _EssentialSkillsCoachState extends State<EssentialSkillsCoach> {
 
   Widget _buildDescription() {
     return Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-      Text(Localization().getStringEx("panel.essential_skills_coach.get_started.description.title", 'Improve your strengths related to:'), style: Styles().textStyles.getTextStyle('panel.essential_skills_coach.header.description'),),
+      Text(Localization().getStringEx("panel.essential_skills_coach.get_started.description.title",
+          'Improve your skills related to:'),
+        style: Styles().textStyles.getTextStyle('panel.essential_skills_coach.header.description'),),
       Padding(padding: EdgeInsets.only(top: 8), child: Text(
         Localization().getStringEx("panel.essential_skills_coach.get_started.description.list", '\t\t\u2022 self-management\n\t\t\u2022 innovation\n\t\t\u2022 cooperation\n\t\t\u2022 social engagement\n\t\t\u2022 emotional resilience'),
         style: Styles().textStyles.getTextStyle('panel.essential_skills_coach.header.description'),
@@ -83,18 +90,34 @@ class _EssentialSkillsCoachState extends State<EssentialSkillsCoach> {
     ]);
   }
 
-  void _onTapStartSkillsCoach() async {
-    //TODO: begin skills coach onboarding, for now create user course and immediately start
-    Future<bool?>? result = AccessDialog.show(context: context, resource: 'academics.essential_skills_coach');
-    if (result == null) {
-      if (StringUtils.isNotEmpty(Config().essentialSkillsCoachKey)) {
-        widget.onStartCourse();
+  // void _onTapStartSkillsCoach() async {
+  //   //TODO: begin skills coach onboarding, for now create user course and immediately start
+  //   Future<bool?>? result = AccessDialog.show(context: context, resource: 'academics.essential_skills_coach');
+  //   if (result == null) {
+  //     if (StringUtils.isNotEmpty(Config().essentialSkillsCoachKey)) {
+  //       widget.onStartCourse?.call(null);
+  //     }
+  //   } else {
+  //     bool? accessResult = await result;
+  //     if (accessResult == true) {
+  //       widget.onStartCourse?.call(null);
+  //     }
+  //   }
+  // }
+
+  void _loadResults() {
+    //_setLoading(true);
+    Surveys().loadUserSurveyResponses(surveyTypes: ["bessi"], limit: 10).then((responses) {
+      if (mounted) {
+        //_setLoading(false);
+        if (CollectionUtils.isNotEmpty(responses)) {
+          responses!.sort(((a, b) => b.dateTaken.compareTo(a.dateTaken)));
+          Navigator.of(context).push(CupertinoPageRoute(builder: (context) => EssentialSkillsResults(latestResponse: responses[0], onStartCourse: widget.onStartCourse)));
+        }
+        else {
+          Navigator.of(context).push(MaterialPageRoute(builder: (context) => EssentialSkillsCoachGetStarted(onStartCourse: widget.onStartCourse)));
+        }
       }
-    } else {
-      bool? accessResult = await result;
-      if (accessResult == true) {
-        widget.onStartCourse();
-      }
-    }
+    });
   }
 }