+++ lib/ui/academics/SkillsSelfEvaluationResultsPanel.dart
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/ui/academics/SkillsSelfEvaluationOccupationListPanel.dart';
 import 'package:illinois/ui/academics/SkillsSelfEvaluation.dart';
@@ -28,6 +29,7 @@ import 'package:rokwire_plugin/ui/popups/popup_message.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 import 'package:rokwire_plugin/ui/widgets/section_header.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 class SkillsSelfEvaluationResultsPanel extends StatefulWidget {
   final SurveyResponse? latestResponse;
@@ -168,6 +170,8 @@ class _SkillsSelfEvaluationResultsPanelState extends State<SkillsSelfEvaluationR
       }
     }
 
+    String contactAddress = Localization().getStringEx('panel.skills_self_evaluation.results.contact.address',
+        'bwrobrts@illinois.edu');
     return [
       Stack(children: [
         responseSections.length > 0 ? ListView.builder(
@@ -219,6 +223,30 @@ class _SkillsSelfEvaluationResultsPanelState extends State<SkillsSelfEvaluationR
         child: Padding(padding: const EdgeInsets.only(top: 4), child: GestureDetector(onTap: _onTapClearAllScores, child:
           Text(Localization().getStringEx('panel.skills_self_evaluation.results.more_info.description', '*Tap score cards for more info'), style: Styles().textStyles.getTextStyle('panel.skills_self_evaluation.content.body.small'), textAlign: TextAlign.left,
         ),)),
+      ),
+      Padding(
+        padding: const EdgeInsets.only(top: 8.0, left: 16.0, right: 16.0),
+        child: RichText(
+          text: TextSpan(children: [
+            TextSpan(text: Localization().getStringEx('panel.skills_self_evaluation.results.contact1',
+              'If you have any questions or concerns about the BESSI score feedback '
+                  'you just received, please contact Dr. Brent Roberts ('),
+            style: Styles().textStyles.getTextStyle('panel.skills_self_evaluation.content.body.small')),
+            TextSpan(text: contactAddress,
+                recognizer: TapGestureRecognizer()..onTap = () {
+                  Uri? uri = Uri.tryParse('mailto:$contactAddress');
+                  if (uri != null) {
+                    launchUrl(uri, mode: LaunchMode.externalApplication);
+                  }
+                },
+                style: Styles().textStyles.getTextStyle('panel.skills_self_evaluation.content.link.small')),
+            TextSpan(text: Localization().getStringEx('panel.skills_self_evaluation.results.contact2',
+                ').  Dr. Roberts is a professor of psychology at the University of '
+                    'Illinois at Urbana-Champaign and a co-creator of the BESSI inventory '
+                    'and associated BESSI assessment tools.'),
+                style: Styles().textStyles.getTextStyle('panel.skills_self_evaluation.content.body.small'))
+          ]),
+        ),
       ),
       Visibility(
         visible: selectedProfile?.params['name'] is String && selectedProfile?.params['definition'] is String,